requirements.txt
aiohttp==3.6.2
aiosqlite==0.15.0
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
chardet==3.0.4
colorama==0.4.3
dateparser==0.7.6
discord.py==1.4.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.2
tzlocal==2.1
wrapt==1.12.1
yarl==1.5.1
